requirements.txt
# Smart-Sync Concierge - Dependencias

# Django Core (Updated for Python 3.9 compatibility)
Django==4.2.27
djangorestframework==3.15.2
django-cors-headers==4.6.0

# IA/ML
qwen==2.5.0
openai==1.59.0  # Para compatibilidad con APIs similares

# Utilidades
python-dateutil==2.9.0
pytz==2024.2
pydantic==2.10.4

# Desarrollo
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
black==24.10.0
ruff==0.8.4

# Seguridad
python-dotenv==1.0.1
